Refactor a bit, moving a class to its own file.
